Reorganize tests files
